// src/app/models/item.model.ts
export interface Item {
    itemName: string;
    quantity: number;
    itemCode: string;
    size: string;
    category: string;
    contact: number;
    pickerName: string;
    pickerSurname: string;
    employeeNumber: string;
    status: string;
    // Add any other properties as needed to match the structure of items in your Firebase collection
  }